filter compoenet seperated from serializers file
